test(core): retry the hook tests' teardown while Windows still holds the directory
`test-windows (core)` failed intermittently in hooks.test.ts with
`EBUSY: resource busy or locked, rmdir`, taking whole runs red on
branches that had not touched hooks. It is not a rerun flake: the
teardown races the operating system, so CI's one retry hits the same
race. It blocked #705, whose diff is the model catalog.

Windows keeps a directory locked while a live process is still using it,
including a script it has not finished exiting from. The five hook
suites write scripts into a temp directory, run them, and remove it as
soon as the assertions pass. The case that failed is the one that runs a
hook to its timeout, so a child outliving the test is exactly what it
arranges. POSIX unlinks a busy directory's entries without complaint,
which is why this is invisible outside CI.

The retry already existed, written for the same reason in the MCP tests
where a stdio server's child holds its working directory. It moves to
test/rm-eventually.ts so both suites share one copy rather than one of
them being fixed and the other left racing. A genuinely stuck directory
still fails, one deadline later, with the error it actually got.
