+++ changelog/unreleased/2026-09-12-windows-hook-temp-lock.zh.md
@@ -1,0 +1,25 @@
+# Windows 上脚本尚未退出时，hook 测试不再失败
+
+- **Date:** 2026-09-12
+- **Type:** process
+- **Scope:** `core`, `ci`
+
+[English](2026-09-12-windows-hook-temp-lock.md)
+
+`test-windows (core)` 会在 `hooks.test.ts` 中间歇性地以 `EBUSY: resource busy or locked, rmdir`
+失败，连带把整次运行判红——哪怕该分支根本没有改动 hook。这里的清理步骤争抢的是操作系统，而不是被测
+代码，因此 CI 已有的那次重试也救不了：第二次运行时同一个竞态依然在。
+
+## 细节
+
+- 只要仍有活着的进程在使用某个目录（包括它尚未退出完毕的脚本），Windows 就会锁住该目录，`rmdir`
+  会返回 `EBUSY`、`EPERM` 或 `ENOTEMPTY`，直到该进程消失为止。五个 hook 测试套件都会把脚本写进一个
+  临时目录、运行它，然后在断言一通过就删除该目录——而此时最后一个子进程未必已经退出。真正失败的那个
+  用例正是把 hook 跑到超时的那个，也就是说「子进程活得比测试久」恰恰是它刻意安排的。
+- POSIX 删除忙碌目录中的条目时不会有任何抱怨，所以这件事在 Linux 与 macOS 上根本看不见。它只出现在
+  CI 上，而且只是偶尔出现。
+- 这五处清理现在改为按时限重试，而不是只删一次。真正卡住的目录仍然会失败，只是晚一个时限，并且带着
+  它实际拿到的那个错误，而不是被吞掉的错误。
+- 这段重试本来就已存在，当初是为同样的原因写在 MCP 测试里的——那里是 stdio server 的子进程占着自己的
+  工作目录。现在它移到 `packages/core/test/rm-eventually.ts`，两套测试共用同一份：手写第二份，正是
+  其中一份被修好、另一份继续踩坑的由来。
