+++ packages/core/test/hooks.test.ts
@@ -2,6 +2,7 @@ import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import { rmEventually } from "./rm-eventually.js";
 import {
   Session,
   assistantText,
@@ -135,7 +136,7 @@ describe("script hooks", () => {
     dir = await fs.mkdtemp(path.join(os.tmpdir(), "penguin-script-hook-"));
   });
   afterEach(async () => {
-    await fs.rm(dir, { recursive: true, force: true });
+    await rmEventually(dir);
   });
 
   const write = async (name: string, body: string): Promise<string> => {
@@ -229,7 +230,7 @@ describe("script pre-tool-use hooks", () => {
     dir = await fs.mkdtemp(path.join(os.tmpdir(), "penguin-tool-script-"));
   });
   afterEach(async () => {
-    await fs.rm(dir, { recursive: true, force: true });
+    await rmEventually(dir);
   });
 
   it("scriptPreToolUseHook feeds the call on stdin and narrows the parsed answer", async () => {
@@ -273,7 +274,7 @@ describe("Session user-prompt hook", () => {
     dir = await fs.mkdtemp(path.join(os.tmpdir(), "penguin-prompt-hook-"));
   });
   afterEach(async () => {
-    await fs.rm(dir, { recursive: true, force: true });
+    await rmEventually(dir);
   });
 
   it("runs the named package's hook with the Session's own id and scratchpad, null when absent", async () => {
@@ -327,7 +328,7 @@ describe("Session stop hooks", () => {
     dir = await fs.mkdtemp(path.join(os.tmpdir(), "penguin-hooks-"));
   });
   afterEach(async () => {
-    await fs.rm(dir, { recursive: true, force: true });
+    await rmEventually(dir);
   });
 
   const fakeEnvironment: EnvironmentInterface = {
@@ -542,7 +543,7 @@ describe("Session pre-tool-use hooks", () => {
     dir = await fs.mkdtemp(path.join(os.tmpdir(), "penguin-tool-hooks-"));
   });
   afterEach(async () => {
-    await fs.rm(dir, { recursive: true, force: true });
+    await rmEventually(dir);
   });
 
   /** Turn 1 calls exec_command; turn 2 answers with the tool output it saw. */
